Extract section refs from sphinx
Do not try to reverse the conversion, it isn't reliable. There are (at
least) few problems:
- real "-" and " " in section title are indistinguishable ("How-to
  guides" -> "how-to-guides")
- some characters are completely removed ("What is Qubes OS?" ->
  "what-is-qubes-os")

Instead, build the site with sphinx and load its refs inventory.
Technically, we need mapping that github-pages used to generate (to map
github-pages links to sphinx refs), but it seems sphinx do it
sufficiently similar.